pyproject.toml
[project]
name = "mcp-tut"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "requests>=2.32.3",
    "python-dotenv==1.0.0"
]
